python/requirements.txt
ultralytics>=8.2.0
requests>=2.31.0
numpy>=1.26.0
Pillow>=10.0.0
